requirements.txt
yfinance
pandas
requests
python-dotenv
kaggle
matplotlib
